feat(corporation): enrich corporation info with home station data from SDE
- Add StationInfo DTO with comprehensive station details
- Integrate SDE service into corporation module for in-memory station lookups
- Enhance CorporationInfo response with home_station field containing:
  - Location hierarchy (region, constellation, solar system)
  - Docking specifications and costs
  - Reprocessing details and security status
- Update module dependencies to pass SDEService through dependency injection
- Fix SDE universe data import issues:
  - Add support for CCP's hierarchical universe structure
  - Fix directory creation during YAML to JSON conversion
  - Make checksum fetching non-blocking to prevent timeout failures
  - Fix StarStatistics temperature field type (int -> float64)

Breaking change: Corporation module constructor now requires SDEService parameter

+++ pkg/sde/types.go
@@ -709,5 +709,5 @@ type StarStatistics struct {
 	Luminosity    float64 `json:"luminosity"`
 	Radius        float64 `json:"radius"`
 	SpectralClass string  `json:"spectralClass"`
-	Temperature   int     `json:"temperature"`
+	Temperature   float64 `json:"temperature"`
 }
